verilog: decoder: fix off by one error in slicing of chunk[0]
For QOI_OP_INDEX(2'b00) looks like we were grabbing more than the
remaining 6 bytes of chunk[0].

This did not yield any bugs since the extra bit was just a 0.

I'm surprised verilator does not raise any warnings or errors about
this (tried even reenabling -WWIDTH). The index variable's index width
should be 5:0, and the old code was trying to store 6:0 into it.

TEST=`make VERILATED=1 test_decode`, hash still matches